feat: updated structure and logic in logout 	    - renamed components to sections in pages folder 	    - add useLocalStorage hook 	    - update server logout logic

+++ server/src/services/auth-service.ts
@@ -62,6 +62,15 @@ export class AuthService {
 	}
 
 	static async logout(token: string) {
+		if (!token) {
+			return
+		}
+
+		const refreshToken = await TokenService.findToken(token);
+		if (!refreshToken) {
+			return
+		}
+
 		await TokenService.removeToken(token);
 	}
